Provide a JsonConverter helper class for converting one type as another. This is especially useful to write objects as simplified data representations, with neglegable overhead.
In this commit it is used to convert common numerics into arrays for serialization and vice versa. This allows a Vector3 to be represented as [X,Y,Z].

Since this will be a common scenario, a sibling JsonConverterTests base class has been implemented in the test repository, for making it easy to test json conversion functionality.