Optimizations, Architectural Improvements and Bug Fixes (#4)
* Fix small errors, add TrimExcess() and TrimExcess(int) to Heatmap2, Heatmap3, and Heatmap3to2.

Add settings to benchmarks on startup. This enables the use of starting up benchmarks from the command line with arguments.

* Add SpanJsonConverter, a powerful and incredibly performant way to serialize data types as arrays of a single unmanaged type, like Vectors and Matrices.

* Rename JsonSerializer to JsonStreamSerializer. But keep JsonSerializer and deprecate it for backwards compatability.

* Deprecate JsonConverter<TInput, TOutput>.

Change HeatmapJsonConverters to inherit from JsonConverter<TInput> and be much more performant.

* Add Counter<TItem>

Counter<TItem> is a wrapper for Dictionary<TItem, int> that makes it simple to count items. For example:
```csharp
public class Program
{
    public static void Main()
    {
        var counter = new Counter<string>();

        counter.Add("hello", 3);
        var i = counter["hello"]; // returns 3

        counter.Add("hello"); // Same as counter.Add("hello", 1);
        i = counter["hello"]; // returns 4

        counter.Remove("hello", 2);
        i = counter["hello"]; // returns 2
    }
}
```

* Change Heatmaps to inherit from Counter<TItem>

This removes a lot of the similar code from both classes and makes the implementation more consistent.

* Add BooleanExtensions

A utility method for turning a boolean into an int / number.

* Update BenchmarkDotNet to version 0.15.1

* Set InvariantGlobalization to true

This ensures that the library is build with invariant globalization in mind, making it function the same way across all regions.

* Change License from GNU to MIT

This change was made so that seperate packages may be build and sold from this software.