Fix Serialization
There were 2 issues to address, 1 with reading and 1 with writing.

Reading: if the offset needed to be moved (due to a negative value in the strengths array), the current position of the offset nor the position of the bounds were taken into account.

Writing: if the very first value was not at the starting position of the heatmap, it would incorrectly move this value one unit to the left and all consecutive units with it. This has been fixed by moving the position after the offset has been copied instead of before.